+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the `opendkim` cookbook will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- ...
+### Removed
+- Metadata: Remove grouping ([RFC-85](https://github.com/chef/chef-rfc/blob/8d47f1d0afa5a2313ed2010e0eda318edc28ba47/rfc085-remove-unused-metadata.md)).
 
 ## [2.0.0] - 2017-03-31
 ### Added in 2.0.0